requirements.txt
Pillow==10.2.0
numpy==1.23.5
gunicorn